Update peakflow routes with user_id

+++ app/Http/Controllers/PeakflowController.php
@@ -7,27 +7,46 @@ use App\Peakflow;
 
 class PeakflowController extends Controller
 {
-    public function index() {
-        return Peakflow::all();
-    }
-
-    // TODO: Return peakflow based on user
-    public function getPeakflowUser($user) {
-        return Peakflow::all();
+    public function index(Request $request) {
+        $user_id = $request->user()->id;
+        return Peakflow::where('user_id', '=', $user_id)->get(); 
     }
 
     public function create(Request $request) {
-        $peakflow = Peakflow::create($request->all());
+        $peakflow = Peakflow::create([
+            'user_id' => $request->user()->id,
+            'date' => request('date'),
+            'time' => request('time'),
+            'measurement_one' => request('measurement_one'),
+            'measurement_two' => request('measurement_two'),
+            'measurement_three' => request('measurement_three'),
+            'taken_medicines' => request('taken_medicines'),
+            'explanation' => request('explanation'),
+        ]);
         return response()->json($peakflow, 201);
     }
 
-    public function update(Request $request, Peakflow $peakflow) {
-        $peakflow->update($request->all());
-        return response()->json($peakflow, 200);
+    public function update(Request $request, $id) {
+        $user_id = $request->user()->id;
+        $peakflow = Peakflow::findOrFail($id);
+
+        if ($peakflow["user_id"] === $user_id) {
+            $peakflow->update($request->all());
+            return response()->json($peakflow, 200);
+        } else {
+            return response()->json("unauthorized", 401);
+        }
     }
 
-    public function delete(Peakflow $peakflow) {
-        $peakflow->delete();
-        return response()->json(null, 204);
+    public function delete(Request $request, $id) {
+        $user_id = $request->user()->id;
+        $peakflow = Peakflow::findOrFail($id);
+
+        if ($peakflow["user_id"] === $user_id) {
+            $peakflow->delete();
+            return response()->json(null, 204);
+        } else {
+            return response()->json("unauthorized", 401);
+        }
     }
 }
